refactor(bar): move out `ShorthandToXamlProperty` from `BarService` to new `XamlHelper` class
