+++ GlazeWM.Bar/BarService.cs
@@ -107,25 +107,5 @@ namespace GlazeWM.Bar
         barWindow.Close();
       });
     }
-
-    /// <summary>
-    /// Convert shorthand properties from user config (ie. `Padding`, `Margin`, and `BorderWidth`)
-    /// to be compatible with their equivalent XAML properties (ie. `Padding`, `Margin`, and
-    /// `BorderThickness`). Shorthand properties follow the 1-to-4 value syntax used in CSS.
-    /// </summary>
-    /// <exception cref="ArgumentException"></exception>
-    public static string ShorthandToXamlProperty(string shorthand)
-    {
-      var shorthandParts = shorthand.Split(" ");
-
-      return shorthandParts.Length switch
-      {
-        1 => shorthand,
-        2 => $"{shorthandParts[1]},{shorthandParts[0]},{shorthandParts[1]},{shorthandParts[0]}",
-        3 => $"{shorthandParts[1]},{shorthandParts[0]},{shorthandParts[1]},{shorthandParts[2]}",
-        4 => $"{shorthandParts[3]},{shorthandParts[0]},{shorthandParts[1]},{shorthandParts[2]}",
-        _ => throw new ArgumentException(null, nameof(shorthand)),
-      };
-    }
   }
 }
